Split xtask verification and benchmarks into modules
verify_system is now a table of labelled remote commands with a
predicate each, so the total is the table length. The benchmark and
metrics analysis move to bench.rs, TestOpts groups its flags into
flattened path, cache, ssh and override structs, and materialize_config
uses one set_if_some helper. The CLI surface is unchanged.
